Move test mixin into a shared location.
