+++ infrastructure/tests/test_commands.py
@@ -1,33 +1,13 @@
-import io
 import json
 import os
 import tempfile
 
-from django.core.management import call_command
 from django.core.management.base import CommandError
 from django.test import TestCase
 
 from locations.models import GeometryStore
+from newsilkroad.test import CommandTestMixin
 from ..models import Initiative, Project
-
-
-class CommandTestMixin(object):
-    """Helpers for running a management command in tests."""
-
-    command = ''
-    defaults = {}
-
-    def call_command(self, *args, **kwargs):
-        command_args = self.defaults.copy()
-        command_args.update(kwargs)
-        stdout = io.StringIO()
-        stderr = io.StringIO()
-        command_args['stdout'] = stdout
-        command_args['stderr'] = stderr
-        call_command(self.command, *args, **command_args)
-        stdout.seek(0)
-        stderr.seek(0)
-        return stdout, stderr
 
 
 class MatchProjectGeometriesTestCase(CommandTestMixin, TestCase):
